requirements.txt
fastapi
uvicorn
transformers
newspaper3k
torch
